Rewrote texture lighting scheme

+++ Library/Engine.h
@@ -121,7 +121,7 @@ private:
 	void handleKeyUp(const SDL_Keycode& code);
 	void handleMouseMotionEvent(const SDL_MouseMotionEvent& event);
 	void illuminateTriangle(Triangle& triangle);
-	void projectTriangle(const Vertex3d (&vertexes)[3], const Vec3 (&worldVecs)[3], const Vec3& normal, const TextureBuffer* texture, int scale);
+	void projectTriangle(const Vertex3d (&vertexes)[3], const Vec3 (&worldVecs)[3], const Vec3& normal, const TextureBuffer* texture, float scale);
 	void update();
 	void updateMovement();
 };
